+++ ChangeLog.md
@@ -17,6 +17,7 @@
   - Fix Unable to change "Icon theme" after disabling it. See: https://github.com/KangLin/RabbitRemoteControl/issues/69
   - Fix the theme to other color when the system theme is dark. See: https://github.com/KangLin/RabbitRemoteControl/pull/68
   - Added Fetching Languages from Environment Variables (LANG). See: https://github.com/KangLin/RabbitCommon/pull/5
+  - Add CTools::LocateFileWithExplorer()
 
 ### Version v2.3.3
   - CAdminAuthorization: Add pkexec in linux